+++ TeamCode/src/main/java/org/betastudio/ftc/thread/TaskMng.java
@@ -2,9 +2,6 @@ package org.betastudio.ftc.thread;
 
 import androidx.annotation.NonNull;
 
-import com.acmerobotics.dashboard.config.ValueProvider;
-
-import org.betastudio.ftc.Interfaces;
 import org.betastudio.ftc.util.Labeler;
 import org.jetbrains.annotations.Contract;
 
@@ -17,32 +14,6 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
 
 public class TaskMng {
-	public static final class TaskFuture implements ValueProvider<String>, Interfaces.ValueProduction <Future<?>> {
-
-		private final String str;
-		private final Future <?> future;
-
-		public TaskFuture(String str, Future <?> future) {
-			this.str = str;
-			this.future = future;
-		}
-
-		@Override
-		public String get() {
-			return str;
-		}
-
-		/**
-		 * @param value 不会干任何事
-		 */
-		@Override
-		public void set(String value) {}
-
-		@Override
-		public Future <?> getVal() {
-			return future;
-		}
-	}
 	@NonNull
 	@Contract(value = "_ -> new", pure = true)
 	public static TaskFuture newTaskFuture(Future <?> future) {
