+++ TeamCode/src/main/java/org/betastudio/ftc/ui/client/implementation/BaseMapClient.java
@@ -4,7 +4,7 @@ import androidx.annotation.NonNull;
 
 import com.acmerobotics.dashboard.config.Config;
 
-import org.betastudio.ftc.thread.TaskMng;
+import org.betastudio.ftc.thread.TaskFuture;
 import org.betastudio.ftc.ui.client.Client;
 import org.betastudio.ftc.ui.client.ClientViewMode;
 import org.betastudio.ftc.ui.client.UpdateConfig;
@@ -171,7 +171,7 @@ public class BaseMapClient implements Client {
 	}
 
 	protected synchronized void updateThreadLines() {
-		for (final TaskMng.TaskFuture task : Global.service.getTasks()) {
+		for (final TaskFuture task : Global.service.getTasks()) {
 			this.telemetry.addData(task.get(), task.getVal().isDone() ? "Done" : "Running");
 		}
 		this.telemetry.update();
